cxx-qt-build: split interface into a separate file
